# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wishify Gifting Marketplace API",
    "description": "Unified REST API for the Wishify Customer and Admin Android applications. Supports flowers, cakes, plants, chocolates, personalized gifts, and combo orders with scheduled delivery slots and Razorpay / COD checkout.",
    "version": "1.0.0",
    "contact": {
      "name": "Wishify Engineering",
      "email": "dev@wishify.in"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.wishify.in",
      "description": "Production Railway Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/products": {
      "get": {
        "summary": "List Products with filters",
        "parameters": [
          { "name": "categoryId", "in": "query", "schema": { "type": "string" } },
          { "name": "occasionId", "in": "query", "schema": { "type": "string" } },
          { "name": "minPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "isVegetarian", "in": "query", "schema": { "type": "boolean" } },
          { "name": "sort", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Paginated product list" }
        }
      }
    },
    "/api/v1/products/{slug}": {
      "get": {
        "summary": "Get Product Detail by Slug",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Product details" }
        }
      }
    },
...
  }
}